Aim of Project
I hope I can change the name of the repository.

The aim of this project is to learn about DAPP, in particular BAT/Brave/Uphold.  The technology, the providers, the users, and its use.

And to help others learn.

This includes a map of the system, with user ratings of DAPP and DAPP resources such as discussion forums, blogs, news, market data, etc.

The users with be rewarded for participating in surveys, and the content and service providers with high ratings and traffic as well.